+++ CHANGELOG.md
@@ -31,6 +31,7 @@ It is recommend to upgrade the storage components first (Receive, Store, etc.) a
 - [#8730](https://github.com/thanos-io/thanos/pull/8730): Receive: add `--remote-write.server-tls-ciphers` to configure cipher suites for the HTTP server.
 - [#8770](https://github.com/thanos-io/thanos/pull/8770): *: add `--grpc-server-tls-curves` to configure curves for gRPC servers.
 - [#8770](https://github.com/thanos-io/thanos/pull/8770): Receive: add `--remote-write.server-tls-curves` to configure curves for the HTTP server.
+- [#8638](https://github.com/thanos-io/thanos/pull/8638): Query: expose Prometheus query timing stats (eval, sort, prep, inner-eval, queue, total) on the `QueryStats` gRPC message. Stats collection is opt-in via the new `enable_stats` flag on `QueryRequest` / `QueryRangeRequest`; pre-existing clients that do not set the flag will no longer receive the `samples_total` / `peak_samples` fields they may have been getting before.
 
 - [#8594](https://github.com/thanos-io/thanos/pull/8594): Query: Support per endpoint TLS configuration